Add support for more long double stuff in glibc.

+++ projects/user-glibc-2.40/math/s_log1pl.c
@@ -1,7 +1,7 @@
 #include <libm-alias-ldouble.h>
 #include <pizlonated_math.h>
 
-int __log1pl (long double x)
+long double __log1pl (long double x)
 {
     return zmath_log1pl (x);
 }
